requirements.txt
Flask
python-dotenv
aiogram
openai
pillow
pytesseract
